+++ Engine/Content/Misc/Layouts/UK-Layout.bin2
@@ -1,0 +1,95 @@
+0000100000101100001
+1000100000101000001
+0000100001001100010
+1000100001001000010
+0000100001101100011
+1000100001101000011
+0000100010001100100
+1000100010001000100
+0000100010101100101
+1000100010101000101
+0000100011001100110
+1000100011001000110
+0000100011101100111
+1000100011101000111
+0000100100001101000
+1000100100001001000
+0000100100101101001
+1000100100101001001
+0000100101001101010
+1000100101001001010
+0000100101101101011
+1000100101101001011
+0000100110001101100
+1000100110001001100
+0000100110101101101
+1000100110101001101
+0000100111001101110
+1000100111001001110
+0000100111101101111
+1000100111101001111
+0000101000001110000
+1000101000001010000
+0000101000101110001
+1000101000101010001
+0000101001001110010
+1000101001001010010
+0000101001101110011
+1000101001101010011
+0000101010001110100
+1000101010001010100
+0000101010101110101
+1000101010101010101
+0000101011001110110
+1000101011001010110
+0000101011101110111
+1000101011101010111
+0000101100001111000
+1000101100001011000
+0000101100101111001
+1000101100101011001
+0000101101001111010
+1000101101001011010
+0001101111101100000
+0000011000100110001
+0000011001000110010
+0000011001100110011
+0000011010000110100
+0000011010100110101
+0000011011000110110
+0000011011100110111
+0000011100000111000
+0000011100100111001
+0000011000000110000
+0001011110100101101
+0001011101100111101
+0001101101101011011
+0001101110101011101
+0001011101000111011
+0001100000000100111
+0001101111000100011
+0001101110001011100
+0001011110000101100
+0001011111000101110
+0001011111100101111
+1000011000100100001
+1000011001000100010
+1000011010000100100
+1000011010100100101
+1000011011001011110
+1000011011100100110
+1000011100000101010
+1000011100100101000
+1000011000000101001
+1001011110101011111
+1001011101100101011
+1001101101101111011
+1001101110101111101
+1001011101000111010
+1001100000001000000
+1001101111001111110
+1001101110001111100
+1001011110000111100
+1001011111000111110
+1001011111100111111
+0111101111001011100
